Clamp and
FRotator GetRelativeRotation() const
Gets the literal value of RelativeRotation. Note, this may be an absolute rotation if this is a root component (not attached to anything) or when GetAbsoluteRotation returns true.
This exists so subclasses don't need to have direct access to the RelativeRotation property so it can be made private later.

CLAMP
template<class T>static T Clamp
(
    const T X,
    const T Min,
    const T Max
)

Remarks
Clamps X to be between Min and Max, inclusive